refactor(api)!: Replace JMS Serializer with symfony Serializer
This change required to touching every controller which forced some
cleanup. Some endpoints were changed in way that technically breaks BC,
but because this API is for now used only privately it's not a big deal.

For the record, changed endpoints:
 - /v1/{provider}/stops/groups -> /v1/{provider}/stop-groups
 - /v1/{provider}/departures/{stop} ->
   /v1/{provider}/stops/{stop}/departures
 - /v1/{provider}/tracks/stops was removed as it duplicated
   /v1/{provider}/track-stops/api/v1/{provider}/track-stops

This changes should make this API more consitent and more aligned with
good practices for REST APIs.