feat(front): Generate sentry config based on browser
This allows to use different ratios for different browser groups based
on the browserslist. This is because we should care more about issues
and replays from current browsers and less for older releases. This is
achieved by multiplying supplied values with multiplier specified for
the groups.

Groups are matched in the release order of browsers, so newer browsers
would be matched first and should have higher sampling rate. The only
exception is for yet unreleased versions of browsers (nightly, canaries
etc.) which we care about, but not as much as for current releases.